requirements.txt
crewai[tools]==0.11.2
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.5.2
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
python-dotenv==1.0.0
openai>=1.12.0
google-search-results==2.4.2
docker>=7.0.0 